fix: detect connection by checking response
This commit fixes the issue where the script could not detect when a connection was established.

The issue was caused by the script using the `adb` command to send a `wget` request to a specific URL.
However, the `wget` command was not returning the HTTP status code in the output.
This made it difficult for the script to determine whether a connection had been established.

To fix this issue, the script has been modified to use the `adb` command to send a `wget` request to a different URL.
This URL returns the HTTP status code in the output, which allows the script to determine whether a connection has been established.