feat: Implement airplane mode feature on smartphone
This commit introduces a new feature that allows the user to enable and disable airplane mode on their smartphone.

This feature is implemented by adding two new functions to the helper/airplane_mode.sh script:

- enable_airplane_mode_on_smartphone
- disable_airplane_mode_on_smartphone

These functions use the adb command to send commands to the smartphone's modem.

This feature will be useful for users who want to quickly and easily enable or disable airplane mode on their smartphone.